correction de bugs

+++ ToDoApp/routers/todos.py
@@ -4,6 +4,8 @@ from fastapi import APIRouter, FastAPI, Depends, HTTPException, Path, dependenci
 import models
 from database import engine, SessionLocal
 from pydantic import BaseModel, Field
+from .auth import get_current_user
+
 
 
 router = APIRouter()
@@ -16,7 +18,9 @@ def get_db():
     finally:
         db.close()
         
+
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -26,31 +30,46 @@ class TodoRequest(BaseModel):
 
         
 @router.get("/", status_code=200)
-async def read_all(db:db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
+
 
 @router.get("/todo/{todo_id}", status_code=200)
-async def read_todo(db:db_dependency, todo_id:int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+           .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
-    raise HTTPException(status_code=404, detail="todos not found")
+    raise HTTPException(status_code=404, detail='Todo not found.')
+
 
 @router.post("/todo", status_code=201)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.dict())
+async def create_todo(user: user_dependency, db: db_dependency,
+                      todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    todo_model = Todos(**todo_request.dict(), owner_id=user.get('id'))
+
     db.add(todo_model)
-    db.commit()  
+    db.commit()
     
     
 @router.put("/todo/{todo_id}")
-async def update_todo(db:db_dependency,
+async def update_todo(user: user_dependency, db: db_dependency,
                       todo_request: TodoRequest,
                       todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id ).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found.')
-    
+
     todo_model.title = todo_request.title
     todo_model.description = todo_request.description
     todo_model.priority = todo_request.priority
@@ -63,10 +82,17 @@ async def update_todo(db:db_dependency,
     
      
 @router.delete("/todo/{todo_id}", status_code=204)
-async def delete_todo(db:db_dependency, todo_id: int = Path(gt=0)):
-        todo_model = db.query(Todos).filter(Todos.id == todo_id ).first()
-        if todo_model is None:
-            raise HTTPException(status_code=404, detail='Todo not found.')
-        
-        db.query(Todos).filter(Todos.id == todo_id).delete()
-        db.commit()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
+    if todo_model is None:
+        raise HTTPException(status_code=404, detail='Todo not found.')
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
+
+    db.commit()
+
+
+
